+++ CPUPipeline/Label.cpp
@@ -1,6 +1,6 @@
 #include "Label.h"
 
-Label::Label(glm::uvec2 posInParent, std::string text, int height, int width, int color)
+Label::Label(glm::vec2 posInParent, std::string text, int height, int width, int color)
 	:Displayable(posInParent), text(text), color(color)
 {
 	setPrefferedSize({ width,height });
