+++ CPUPipeline/Panel.cpp
@@ -1,6 +1,6 @@
 #include "Panel.h"
 
-Panel::Panel(int color, glm::uvec2 posInParent, int width, int height)
+Panel::Panel(int color, glm::vec2 posInParent, int width, int height)
 	:Displayable(posInParent), color(color)
 {
 	setPrefferedSize({ width, height });
@@ -11,7 +11,7 @@ void Panel::onPaint(GUIUtils& guiUtils, int x, int y, int width, int height) con
 	guiUtils.fb.DrawRect(x, y, x + width, y + height, color);
 	for (Displayable* child : children)
 	{
-		glm::uvec2 childPosInParent = child->getPosInParent();
+		glm::vec2 childPosInParent = child->getPosInParent();
 		child->Display(guiUtils, x + childPosInParent.x, y + childPosInParent.y, width, height);
 	}
 }
